#3 🐘 rename folder notFoundPage -> notFound
