+++ src/notFound/urlSuggestions/getRestUrlSuggestions/getPatternRestUrlMeaningfulString.test.ts
