+++ src/notFound/urlSuggestions/getRestUrlSuggestions/getPatternRestUrlMeaningfulString.ts
